Image collage: Allow image to expand to 100% width
This makes it feel more complete, the image remains responsive whilst
filling the screen.  It may take up substantially more real-estate on
larger monitors but it also doesn't make it feel out of place.

The only other recommendation here would be to add several more images
to the collage allowing the width to be far more expanded.  Then setting
the width to 100% but having the overflow hidden.